Cambio atributos y métodos

+++ Ritmubi/src/com/ritmubi/Ciclo.java
@@ -2,26 +2,33 @@ package com.ritmubi;
 
 public class Ciclo {
 
-	private short 
+	private float 
 		emocional,
 		fisico,
 		intelectual;
-	
-	public Ciclo(short emocional, short fisico, short intelectual) {
-		this.emocional = emocional;
-		this.fisico = fisico;
-		this.intelectual = intelectual;
+
+	/**
+	 * Crea una nueva instancia de la clase Ciclo.
+	 * Un Objeto de tipo ciclo contiene los estados emocional, fisico e intelectual  
+	 * @param emocional
+	 * @param fisico
+	 * @param intelectual
+	 */
+	public Ciclo(double emocional, double fisico, double intelectual) {
+		this.emocional = (float) emocional;
+		this.fisico = (float) fisico;
+		this.intelectual = (float) intelectual;
 	}
 
-	public short getEmocional() {
+	public float getEmocional() {
 		return emocional;
 	}
 
-	public short getFisico() {
+	public float getFisico() {
 		return fisico;
 	}
 
-	public short getIntelectual() {
+	public float getIntelectual() {
 		return intelectual;
 	}
 }
